% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstars.R
\name{pstar}
\alias{pstar}
\title{Convert numeric pvalues to character vector.}
\usage{
pstar(pvalue)
}
\arguments{
\item{pvalues}{vector of numeric pvalues to be converted}
}
\value{
A character vector of same length as input
}
\description{
\code{pstars} converts numeric pvalues to a character vector of asterisk symbols.
}
\examples{
pstars(c(0.04, 0.005, 0.1))
}

