% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstars.R
\name{pstar}
\alias{pstar}
\title{Wrapper functions to convert gene identifiers in org.Hs.eg.db or org.Mm.eg.db package}
\usage{
pstar(pvalue)
}
\arguments{
\item{pvalues}{vector of numeric pvalues to be converted}
}
\value{
A character vector of same length as input
}
\description{
\code{pstars} converts ids from identifier X to identifier Y using AnnotationDbi.
By default, target ids are concatenated for non-unique mappings, but non-unique mappings can be
forced to return unique mappings. However, this is not recommended.
}
\examples{
pstars(c(0.04, 0.005, 0.1))
}

